S4 class for assignment parameters

An S4 class to store assignment parameters.

Slots

technique

the analytical technique

correlations_parameters

a list of correlation parameters to be passed to metabolyseR::correlations()

max_M

the maximum molecular mass for which to assign molecular formulas

MF_rank_threshold

rank threshold for molecular formula selection

ppm

the parts per million error threshold

limit

the atomic mass unit deviation limit for relationship calculation

RT_diff_limit

the limit for retention time differences for correlated features in adduct and isotopic assignment

adducts

a list of character vectors containing the adducts names. List element names should denote ionisation mode. The order that these adducts are provided denotes their expected relative importance to assignments with the first expected to be the most common and the last the least common within each ionisation mode.

isotopes

a character vector of isotopes to use. Similarly to the adducts, their order denotes the expected commonality in the data.

transformations

a character vector of transformations molecular formula changes

adduct_rules

a tibble containing the adduct formation rules as returned by mzAnnotation::adduct_rules()

isotope_rules

a tibble containing the isotope rules as returned by mzAnnotation::isotope_rules()

transformation_rules

tibble containing the transformation rules as returned by mzAnnotation::transformation_rules()