Assignment accessors

Access methods for Assignment S4 class

Usage

featureData(assignment)

# S4 method for Assignment
featureData(assignment)

correlations(assignment)

# S4 method for Assignment
correlations(assignment)

relationships(assignment)

# S4 method for Assignment
relationships(assignment)

iterations(assignment)

# S4 method for Assignment
iterations(assignment)

graph(assignment, iteration, type = c("selected", "all"))

# S4 method for Assignment
graph(assignment, iteration, type = c("selected", "all"))

components(assignment, iteration, type = c("selected", "all"))

# S4 method for Assignment
components(assignment, iteration, type = c("selected", "all"))

featureComponents(assignment, feature, type = c("selected", "all"))

# S4 method for Assignment
featureComponents(assignment, feature, type = c("selected", "all"))

component(assignment, component, iteration, type = c("selected", "all"))

# S4 method for Assignment
component(assignment, component, iteration, type = c("selected", "all"))

assignments(assignment)

# S4 method for Assignment
assignments(assignment)

assignedData(assignment)

# S4 method for Assignment
assignedData(assignment)

summariseAssignments(assignment)

# S4 method for Assignment
summariseAssignments(assignment)

Arguments

assignment

S4 object of class Assignment

iteration

the assignment iteration

type

the graph type to return. filtered returns the assignment graph after component selection. all returns all assignment components.

feature

feature information to extract

component

component number to extract

Value

A tibble or tbl_graph containing assignment results depending on the method used.

Details

featureData - Return the initially specifed m/z feature data.

correlations - Return the correlation analysis results.

relationships - Return the calculated relationships.

iterations - Return the assignment iteration performed.

graph - Return a selected graph.

components - Return the component information for an assignment iteration.

featureComponents - Return the component information for a selected feature.

component - Extract a component graph.

assignments - Return the molecular formulas assigned to the m/z features.

assignedData - Return the m/z peak intensity matrix with the molecular formula assignments included in the column names.

summariseAssignments - Return a tibble of the assignments summarised by molecular formula.

Examples

plan(future::sequential)
p <- assignmentParameters('FIE-HRMS')

mf_assignments <- assignMFs(feature_data,p)
#> 
#> assignments v1.0.0 Tue Feb 21 15:44:36 2023
#> ________________________________________________________________________________
#> Assignment Parameters: 
#> 
#> 	 Technique:		 FIE-HRMS 
#> 	 Max M:			 800 
#> 	 MF rank threshold:	 3 
#> 	 PPM threshold:		 6 
#> 	 Relationship limit:	 0.001 
#> 	 RT limit:		  
#> 	 Correlations:
#> 		method: spearman
#>  		pAdjustMethod: bonferroni
#>  		corPvalue: 0.05
#>  		minCoef: 0.7
#>  		maxCor: Inf
#> 
#> 	 Adducts: 
#> 	 n: [M-H]1-, [M+Cl]1-, [M+K-2H]1-, [M-2H]2-, [M+Cl37]1-, [2M-H]1-
#> 	 p: [M+H]1+, [M+K]1+, [M+Na]1+, [M+K41]1+, [M+2H]2+, [2M+H]1+ 
#> 	 Isotopes: 13C, 18O, 13C2 
#> 	 Transformations: M - [O] + [NH2], M - [OH] + [NH2], M + [H2], M - [H2] + [O], M - [H] + [CH3], M - [H] + [NH2], M - [H] + [OH], M + [H2O], M - [H3] + [H2O], M - [H] + [CHO2], M - [H] + [SO3], M - [H] + [PO3H2]
#> ________________________________________________________________________________
#> No. m/z:	10
#> Calculating correlations …
#> Calculating correlations 	✔ [10 correlations] [0.5S]
#> Calculating relationships …
#> Calculating relationships 	✔ [3.7S]
#> Adduct & isotopic assignment …
#> generating molecular formulas…
#> generating molecular formulas	✔ [39.9S]
#> iteration 1…
#> iteration 1			✔ [1S]
#> iteration 2…
#> Adduct & isotopic assignment 	✔ [42.7S]
#> Transformation assignment…
#> iteration 1 …
#> iteration 1 			✔ [2.3S]
#> iteration 2 …
#> Transformation assignment 	✔ [2.4S]
#> ________________________________________________________________________________
#> 
#> Complete! [49.8S]

## Return feature data
featureData(mf_assignments)
#> # A tibble: 60 × 10
#>    n191.01962 n192.02306 n193.…¹ n226.…² n228.…³ n228.…⁴ n231.…⁵ n384.…⁶ n385.…⁷
#>         <dbl>      <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1     0.109     0.00401 6.09e-4 4.17e-4 1.35e-3 9.12e-5 1.22e-3 3.11e-4 3.10e-5
#>  2     0.0955    0.00346 4.66e-4 3.06e-4 1.10e-3 8.47e-5 8.31e-4 1.00e-4 8.88e-7
#>  3     0.109     0.00418 6.37e-4 2.94e-4 1.13e-3 8.79e-5 1.26e-3 3.64e-4 8.01e-5
#>  4     0.102     0.00360 5.16e-4 3.66e-4 1.07e-3 1.07e-4 1.13e-3 2.06e-4 6.73e-5
#>  5     0.0956    0.00315 5.20e-4 4.87e-4 8.73e-4 1.11e-4 8.40e-4 1.72e-4 0      
#>  6     0.0940    0.00350 4.40e-4 3.11e-4 1.12e-3 2.28e-5 1.06e-3 1.52e-4 4.97e-5
#>  7     0.128     0.00578 7.84e-4 3.33e-4 1.13e-3 1.37e-4 1.40e-3 5.30e-4 1.49e-4
#>  8     0.0890    0.00345 4.83e-4 3.65e-4 1.04e-3 1.00e-4 9.51e-4 1.89e-4 8.42e-5
#>  9     0.0887    0.00317 4.69e-4 4.63e-4 8.29e-4 1.51e-4 9.22e-4 2.25e-4 4.32e-5
#> 10     0.0996    0.00369 5.39e-4 4.90e-4 8.66e-4 1.93e-4 1.00e-3 2.55e-4 1.63e-5
#> # … with 50 more rows, 1 more variable: p208.04518 <dbl>, and abbreviated
#> #   variable names ¹​n193.02388, ²​n226.99693, ³​n228.97636, ⁴​n228.99274,
#> #   ⁵​n231.00069, ⁶​n384.0495, ⁷​n385.04874

## Return correlations
correlations(mf_assignments)
#> # A tibble: 10 × 12
#>    Feature1   Feature2 Mode1 Mode2 `m/z1` `m/z2` Reten…¹ Reten…² Reten…³ log2I…⁴
#>    <chr>      <chr>    <chr> <chr>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 n191.01962 p208.04… n     p       191.   208.      NA      NA      NA    0   
#>  2 n191.01962 n192.02… n     n       191.   192.      NA      NA      NA    4.80
#>  3 n192.02306 p208.04… n     p       192.   208.      NA      NA      NA   -4.80
#>  4 n231.00069 n384.04… n     n       231.   384.      NA      NA      NA    2.36
#>  5 n226.99693 n228.99… n     n       227.   229.      NA      NA      NA    1.74
#>  6 n191.01962 n193.02… n     n       191.   193.      NA      NA      NA    7.56
#>  7 n193.02388 p208.04… n     p       193.   208.      NA      NA      NA   -7.56
#>  8 n192.02306 n193.02… n     n       192.   193.      NA      NA      NA    2.76
#>  9 n191.01962 n231.00… n     n       191.   231.      NA      NA      NA    6.96
#> 10 n231.00069 p208.04… n     p       231.   208.      NA      NA      NA   -6.96
#> # … with 2 more variables: coefficient <dbl>, ID <int>, and abbreviated
#> #   variable names ¹​RetentionTime1, ²​RetentionTime2, ³​RetentionTimeDiff,
#> #   ⁴​log2IntensityRatio

## Return relationships
relationships(mf_assignments)
#> # A tibble: 52 × 19
#>    Feature1   Feature2 Mode1 Mode2 `m/z1` `m/z2` Reten…¹ Reten…² Reten…³ Adduct1
#>    <chr>      <chr>    <chr> <chr>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl> <chr>  
#>  1 n191.01962 p208.04… n     p       191.   208.      NA      NA      NA [M-H]1-
#>  2 n191.01962 p208.04… n     p       191.   208.      NA      NA      NA [M-H]1-
#>  3 n191.01962 p208.04… n     p       191.   208.      NA      NA      NA [M-H]1-
#>  4 n191.01962 p208.04… n     p       191.   208.      NA      NA      NA [M-H]1-
#>  5 n191.01962 p208.04… n     p       191.   208.      NA      NA      NA [M+K-2…
#>  6 n191.01962 p208.04… n     p       191.   208.      NA      NA      NA [M+K-2…
#>  7 n191.01962 p208.04… n     p       191.   208.      NA      NA      NA [M+K-2…
#>  8 n191.01962 p208.04… n     p       191.   208.      NA      NA      NA [M+K-2…
#>  9 n191.01962 p208.04… n     p       191.   208.      NA      NA      NA [M+K-2…
#> 10 n191.01962 p208.04… n     p       191.   208.      NA      NA      NA [M+K-2…
#> # … with 42 more rows, 9 more variables: Adduct2 <chr>, Isotope1 <chr>,
#> #   Isotope2 <chr>, Transformation1 <chr>, Transformation2 <chr>,
#> #   log2IntensityRatio <dbl>, coefficient <dbl>, Error <dbl>, ID <int>, and
#> #   abbreviated variable names ¹​RetentionTime1, ²​RetentionTime2,
#> #   ³​RetentionTimeDiff

## Return the available iterations
iterations(mf_assignments)
#> [1] "A&I1" "T1"  

## Return a selected graph
graph(mf_assignments,'A&I1')
#> # A tbl_graph: 5 nodes and 4 edges
#> #
#> # An undirected simple graph with 2 components
#> #
#> # Node Data: 5 × 21 (active)
#>   name  Feature Retent… Isotope Adduct MF    Theore… Measur… Theore… Measur…
#>   <chr> <chr>     <dbl> <chr>   <chr>  <chr>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 n191… n191.0…      NA NA      [M-H]… C6H8…    192.    192.    191.    191.
#> 2 n192… n192.0…      NA 13C     [M-H]… C6H8…    192.    192.    192.    192.
#> 3 n193… n193.0…      NA 18O     [M-H]… C6H8…    192.    192.    193.    193.
#> 4 n226… n226.9…      NA NA      [M+Cl… C6H8…    192.    192.    227.    227.
#> 5 n228… n228.9…      NA NA      [M+Cl… C6H8…    192.    192.    229.    229.
#> # … with 11 more variables: `PPM error` <dbl>, `MF Plausibility (%)` <dbl>,
#> #   AIS <dbl>, ID <int>, Component <dbl>, Size <dbl>, Nodes <int>,
#> #   Degree <dbl>, Density <dbl>, `Component Plausibility` <dbl>, Weight <dbl>
#> #
#> # Edge Data: 4 × 23
#>    from    to Featur… Featur… Mode1 Mode2 `m/z1` `m/z2` Retent… Retent… Retent…
#>   <int> <int> <chr>   <chr>   <chr> <chr>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>
#> 1     1     2 n191.0… n192.0… n     n       191.   192.      NA      NA      NA
#> 2     1     3 n191.0… n193.0… n     n       191.   193.      NA      NA      NA
#> 3     2     3 n192.0… n193.0… n     n       192.   193.      NA      NA      NA
#> # … with 1 more row, and 12 more variables: Adduct1 <chr>, Adduct2 <chr>,
#> #   Isotope1 <chr>, Isotope2 <chr>, Transformation1 <chr>,
#> #   Transformation2 <chr>, log2IntensityRatio <dbl>, coefficient <dbl>,
#> #   Error <dbl>, ID <int>, MF1 <chr>, MF2 <chr>

## Return a component information for a selected graph
components(mf_assignments,'A&I1')
#> # A tibble: 2 × 6
#>   Component  Size Nodes Degree Density `Component Plausibility`
#>       <dbl> <dbl> <int>  <dbl>   <dbl>                    <dbl>
#> 1         1     3     3      2       1                0.000346 
#> 2        24     1     2      1       1                0.0000272

## Return a component information for a selected feature
featureComponents(mf_assignments,'n191.01962')
#> # A tibble: 2 × 22
#>   Iteration name    Feature Reten…¹ Isotope Adduct MF    Theor…² Measu…³ Theor…⁴
#>   <chr>     <chr>   <chr>     <dbl> <chr>   <chr>  <chr>   <dbl>   <dbl>   <dbl>
#> 1 A&I1      n191.0… n191.0…      NA NA      [M-H]… C6H8…    192.    192.    191.
#> 2 T1        n191.0… n191.0…      NA NA      [M-H]… C6H8…    192.    192.    191.
#> # … with 12 more variables: `Measured m/z` <dbl>, `PPM error` <dbl>,
#> #   `MF Plausibility (%)` <dbl>, AIS <dbl>, ID <int>, Component <dbl>,
#> #   Size <dbl>, Nodes <int>, Degree <dbl>, Density <dbl>,
#> #   `Component Plausibility` <dbl>, Weight <dbl>, and abbreviated variable
#> #   names ¹​RetentionTime, ²​`Theoretical M`, ³​`Measured M`, ⁴​`Theoretical m/z`

 ## Extract a component graph
component(mf_assignments,1,'A&I1')
#> # A tbl_graph: 3 nodes and 3 edges
#> #
#> # An undirected simple graph with 1 component
#> #
#> # Node Data: 3 × 21 (active)
#>   name  Feature Retent… Isotope Adduct MF    Theore… Measur… Theore… Measur…
#>   <chr> <chr>     <dbl> <chr>   <chr>  <chr>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 n191… n191.0…      NA NA      [M-H]… C6H8…    192.    192.    191.    191.
#> 2 n192… n192.0…      NA 13C     [M-H]… C6H8…    192.    192.    192.    192.
#> 3 n193… n193.0…      NA 18O     [M-H]… C6H8…    192.    192.    193.    193.
#> # … with 11 more variables: `PPM error` <dbl>, `MF Plausibility (%)` <dbl>,
#> #   AIS <dbl>, ID <int>, Component <dbl>, Size <dbl>, Nodes <int>,
#> #   Degree <dbl>, Density <dbl>, `Component Plausibility` <dbl>, Weight <dbl>
#> #
#> # Edge Data: 3 × 23
#>    from    to Featur… Featur… Mode1 Mode2 `m/z1` `m/z2` Retent… Retent… Retent…
#>   <int> <int> <chr>   <chr>   <chr> <chr>  <dbl>  <dbl>   <dbl>   <dbl>   <dbl>
#> 1     1     2 n191.0… n192.0… n     n       191.   192.      NA      NA      NA
#> 2     1     3 n191.0… n193.0… n     n       191.   193.      NA      NA      NA
#> 3     2     3 n192.0… n193.0… n     n       192.   193.      NA      NA      NA
#> # … with 12 more variables: Adduct1 <chr>, Adduct2 <chr>, Isotope1 <chr>,
#> #   Isotope2 <chr>, Transformation1 <chr>, Transformation2 <chr>,
#> #   log2IntensityRatio <dbl>, coefficient <dbl>, Error <dbl>, ID <int>,
#> #   MF1 <chr>, MF2 <chr>

## Return assignments
assignments(mf_assignments)
#> # A tibble: 6 × 15
#>   Name      Feature Reten…¹ Isotope Adduct MF    Theor…² Measu…³ Theor…⁴ Measu…⁵
#>   <chr>     <chr>     <dbl> <chr>   <chr>  <chr>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1 n191.019… n191.0…      NA NA      [M-H]… C6H8…    192.    192.    191.    191.
#> 2 n192.023… n192.0…      NA 13C     [M-H]… C6H8…    192.    192.    192.    192.
#> 3 n193.023… n193.0…      NA 18O     [M-H]… C6H8…    192.    192.    193.    193.
#> 4 n226.996… n226.9…      NA NA      [M+Cl… C6H8…    192.    192.    227.    227.
#> 5 n228.992… n228.9…      NA NA      [M+Cl… C6H8…    192.    192.    229.    229.
#> 6 p208.045… p208.0…      NA NA      [M+H]… C6H9…    207.    207.    208.    208.
#> # … with 5 more variables: `PPM error` <dbl>, `MF Plausibility (%)` <dbl>,
#> #   Mode <chr>, Component <dbl>, Iteration <chr>, and abbreviated variable
#> #   names ¹​RetentionTime, ²​`Theoretical M`, ³​`Measured M`, ⁴​`Theoretical m/z`,
#> #   ⁵​`Measured m/z`

## Return an m/z intensity matrix with the assignments included
## in the column names
assignedData(mf_assignments)
#> # A tibble: 60 × 10
#>    n191.01962 …¹ n192.…² n193.…³ n226.…⁴ n228.…⁵ n228.…⁶ n231.…⁷ n384.…⁸ n385.…⁹
#>            <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1        0.109  0.00401 6.09e-4 4.17e-4 1.35e-3 9.12e-5 1.22e-3 3.11e-4 3.10e-5
#>  2        0.0955 0.00346 4.66e-4 3.06e-4 1.10e-3 8.47e-5 8.31e-4 1.00e-4 8.88e-7
#>  3        0.109  0.00418 6.37e-4 2.94e-4 1.13e-3 8.79e-5 1.26e-3 3.64e-4 8.01e-5
#>  4        0.102  0.00360 5.16e-4 3.66e-4 1.07e-3 1.07e-4 1.13e-3 2.06e-4 6.73e-5
#>  5        0.0956 0.00315 5.20e-4 4.87e-4 8.73e-4 1.11e-4 8.40e-4 1.72e-4 0      
#>  6        0.0940 0.00350 4.40e-4 3.11e-4 1.12e-3 2.28e-5 1.06e-3 1.52e-4 4.97e-5
#>  7        0.128  0.00578 7.84e-4 3.33e-4 1.13e-3 1.37e-4 1.40e-3 5.30e-4 1.49e-4
#>  8        0.0890 0.00345 4.83e-4 3.65e-4 1.04e-3 1.00e-4 9.51e-4 1.89e-4 8.42e-5
#>  9        0.0887 0.00317 4.69e-4 4.63e-4 8.29e-4 1.51e-4 9.22e-4 2.25e-4 4.32e-5
#> 10        0.0996 0.00369 5.39e-4 4.90e-4 8.66e-4 1.93e-4 1.00e-3 2.55e-4 1.63e-5
#> # … with 50 more rows, 1 more variable: `p208.04518 C6H9NO7  [M+H]1+` <dbl>,
#> #   and abbreviated variable names ¹​`n191.01962 C6H8O7  [M-H]1-`,
#> #   ²​`n192.02306 C6H8O7 13C [M-H]1-`, ³​`n193.02388 C6H8O7 18O [M-H]1-`,
#> #   ⁴​`n226.99693 C6H8O7  [M+Cl]1-`, ⁵​n228.97636,
#> #   ⁶​`n228.99274 C6H8O7  [M+Cl37]1-`, ⁷​n231.00069, ⁸​n384.0495, ⁹​n385.04874

## Return the assignments summarised by molecular formula
summariseAssignments(mf_assignments)
#> # A tibble: 2 × 4
#>   MF      Features                                                 Isoto…¹ Count
#>   <chr>   <chr>                                                    <chr>   <int>
#> 1 C6H8O7  n191.01962; n192.02306; n193.02388; n226.99693; n228.99… " [M-H…     5
#> 2 C6H9NO7 p208.04518                                               " [M+H…     1
#> # … with abbreviated variable name ¹​`Isotopes & Ionisation Products`